+++ test/cli/commands/smoke-test/smoke-tests/aws/resource-tests/vpc-smoke-tests.test.ts
@@ -135,32 +135,15 @@ describe('vpc smoke tests', () => {
         "format": IacFormat.tf,
         "resourceType": "aws_vpc",
         "changeType": ChangeType.CREATE,
-        "resourceRecord": {
-          "address": "module.ts_aws_vpc_hello_world.aws_vpc.ts_aws_vpc",
-          "type": "aws_vpc",
-          "logicalId": "ts_aws_vpc",
-          "tfProviderName": "registry.terraform.io/hashicorp/aws",
-          "tfReferences": [
-            {
-              "property": "cidr_block",
-              "reference": "module.ts_aws_vpc_hello_world.var.ts_aws_vpc_cidr_block"
-            }
-          ],
-          "properties": {
-            "assign_generated_ipv6_cidr_block": null,
-            "cidr_block": "10.0.0.0/16",
-            "enable_dns_hostnames": true,
-            "enable_dns_support": true,
-            "instance_tenancy": "default",
-            "ipv4_ipam_pool_id": null,
-            "ipv4_netmask_length": null,
-            "ipv6_ipam_pool_id": null,
-            "ipv6_netmask_length": null,
-            "tags": null
-          }
+        "address": "module.ts_aws_vpc_hello_world.aws_vpc.ts_aws_vpc",
+        "logicalId": "ts_aws_vpc",
+        "providerName": "registry.terraform.io/hashicorp/aws",
+        "properties": {
+          "cidrBlock": "10.0.0.0/16",
+          "instanceTenancy": "default",
         }
       };
-      const allResources = require('../../../test-data/vpc-stack/TfDiffWithNat.json');
+      const allResources = require('../../../test-data/vpc-stack/tf/with-nat/tf-diff.json');
 
       let thrownError;
       try {
@@ -174,35 +157,24 @@ describe('vpc smoke tests', () => {
     });
     it('does nothing if requirePrivateSubnet is false', async () => {
       const mockVpc: ResourceDiffRecord = {
-        "format": IacFormat.tf,
-        "resourceType": "aws_vpc",
+        "stackName": "SmokeTestApp",
+        "format": IacFormat.awsCdk,
         "changeType": ChangeType.CREATE,
-        "resourceRecord": {
-          "address": "module.ts_aws_vpc_hello_world.aws_vpc.ts_aws_vpc",
-          "type": "aws_vpc",
-          "logicalId": "ts_aws_vpc",
-          "tfProviderName": "registry.terraform.io/hashicorp/aws",
-          "tfReferences": [
+        "resourceType": "AWS::EC2::VPC",
+        "address": "Vpc/Vpc",
+        "logicalId": "VpcC3027511",
+        "properties": {
+          "cidrBlock": "10.40.0.0/16",
+          "instanceTenancy": "default",
+          "tagSet": [
             {
-              "property": "cidr_block",
-              "reference": "module.ts_aws_vpc_hello_world.var.ts_aws_vpc_cidr_block"
+              "Key": "Name",
+              "Value": "SmokeTestApp/Vpc/Vpc"
             }
-          ],
-          "properties": {
-            "assign_generated_ipv6_cidr_block": null,
-            "cidr_block": "10.0.0.0/16",
-            "enable_dns_hostnames": true,
-            "enable_dns_support": true,
-            "instance_tenancy": "default",
-            "ipv4_ipam_pool_id": null,
-            "ipv4_netmask_length": null,
-            "ipv6_ipam_pool_id": null,
-            "ipv6_netmask_length": null,
-            "tags": null
-          }
+          ]
         }
       };
-      const allResources = require('../../../test-data/vpc-stack/TfDiffWithNat.json');
+      const allResources = require('../../../test-data/vpc-stack/tf/with-nat/tf-diff.json');
 
       let thrownError;
       try {
@@ -214,177 +186,73 @@ describe('vpc smoke tests', () => {
         expect(thrownError).toBeUndefined();
       }
     });
-    describe('terraform', () => {
-      it('validates the route table for the private subnet has a route pointing to a NAT gateway', async () => {
-        const mockVpc = {
-          "format": "tf",
-          "resourceType": "aws_vpc",
-          "changeType": "CREATE",
-          "resourceRecord": {
-            "address": "module.ts_aws_vpc_hello_world.aws_vpc.ts_aws_vpc",
-            "type": "aws_vpc",
-            "logicalId": "ts_aws_vpc",
-            "tfProviderName": "registry.terraform.io/hashicorp/aws",
-            "tfReferences": [
-              {
-                "property": "cidr_block",
-                "reference": "module.ts_aws_vpc_hello_world.var.ts_aws_vpc_cidr_block"
-              }
-            ],
-            "properties": {
-              "assign_generated_ipv6_cidr_block": null,
-              "cidr_block": "10.0.0.0/16",
-              "enable_dns_hostnames": true,
-              "enable_dns_support": true,
-              "instance_tenancy": "default",
-              "ipv4_ipam_pool_id": null,
-              "ipv4_netmask_length": null,
-              "ipv6_ipam_pool_id": null,
-              "ipv6_netmask_length": null,
-              "tags": null
+    it('validates the route table for the private subnet has a route pointing to a NAT gateway', async () => {
+      const mockVpc = {
+        "stackName": "SmokeTestApp",
+        "format": "aws-cdk",
+        "changeType": "CREATE",
+        "resourceType": "AWS::EC2::VPC",
+        "address": "Vpc/Vpc",
+        "logicalId": "VpcC3027511",
+        "properties": {
+          "cidrBlock": "10.40.0.0/16",
+          "instanceTenancy": "default",
+          "tagSet": [
+            {
+              "Key": "Name",
+              "Value": "SmokeTestApp/Vpc/Vpc"
             }
-          }
-        } as ResourceDiffRecord;
-        const allResources = require('../../../test-data/vpc-stack/TfDiffWithNat.json');
-  
-        let thrownError;
-        try {
-          await vpcSmokeTest(mockVpc, allResources, { requirePrivateSubnet: true });
-        } catch (error) {
-          thrownError = error;
-        } finally {
-          expect(mockLoggerInfo).toBeCalled();
-          expect(mockLoggerInfo).toBeCalledWith('Verifying subnet configuration...');
-          expect(thrownError).toBeUndefined();
+          ]
         }
-      });
-      it('throws if there is not a subnet with a route pointing to a NAT gateway', async () => {
-        const mockVpc: ResourceDiffRecord = {
-          "format": IacFormat.tf,
-          "resourceType": "aws_vpc",
-          "changeType": ChangeType.CREATE,
-          "resourceRecord": {
-            "address": "aws_vpc.ts_aws_vpc",
-            "type": "aws_vpc",
-            "logicalId": "ts_aws_vpc",
-            "tfProviderName": "registry.terraform.io/hashicorp/aws",
-            "tfReferences": [
-              {
-                "property": "cidr_block",
-                "reference": "var.ts_aws_vpc_cidr_block"
-              }
-            ],
-            "properties": {
-              "assign_generated_ipv6_cidr_block": null,
-              "cidr_block": "10.0.0.0/16",
-              "enable_dns_hostnames": true,
-              "enable_dns_support": true,
-              "instance_tenancy": "default",
-              "ipv4_ipam_pool_id": null,
-              "ipv4_netmask_length": null,
-              "ipv6_ipam_pool_id": null,
-              "ipv6_netmask_length": null,
-              "tags": null
-            }
-          }
-        };
-        const allResources = require('../../../test-data/vpc-stack/TfDiffNoNat.json');
-  
-        let thrownError;
-        try {
-          await vpcSmokeTest(mockVpc, allResources, { requirePrivateSubnet: true });
-        } catch (error) {
-          thrownError = error;
-        } finally {
-          expect(mockLoggerInfo).toBeCalled();
-          expect(mockLoggerInfo).toBeCalledWith('Verifying subnet configuration...');
-          
-          expect(thrownError).toBeDefined();
-          expect(thrownError).toHaveProperty('name', 'CustomError');
-          expect(thrownError).toHaveProperty('message', 'Missing private subnets!');
-          expect(thrownError).toHaveProperty('reason', 'Based on the configuration passed, private subnets with a NAT Gateway are required for all vpcs but none was found for "ts_aws_vpc".');
-        }
-      });
+      } as ResourceDiffRecord;
+      const allResources = require('../../../test-data/vpc-stack/cdk/with-nat/aws-cdk-diff.json');
+
+      let thrownError;
+      try {
+        await vpcSmokeTest(mockVpc, allResources, { requirePrivateSubnet: true });
+      } catch (error) {
+        thrownError = error;
+      } finally {
+        expect(mockLoggerInfo).toBeCalled();
+        expect(mockLoggerInfo).toBeCalledWith('Verifying subnet configuration...');
+        expect(thrownError).toBeUndefined();
+      }
     });
-    describe('aws-cdk', () => {
-      it('validates the route table for the private subnet has a route pointing to a NAT gateway', async () => {
-        const mockVpc = {
-          "stackName": "SmokeTestApp",
-          "format": "aws-cdk",
-          "changeType": "CREATE",
-          "resourceType": "AWS::EC2::VPC",
-          "resourceRecord": {
-            "address": "Vpc/Vpc",
-            "type": "AWS::EC2::VPC",
-            "logicalId": "VpcC3027511",
-            "properties": {
-              "CidrBlock": "10.40.0.0/16",
-              "EnableDnsHostnames": true,
-              "EnableDnsSupport": true,
-              "InstanceTenancy": "default",
-              "Tags": [
-                {
-                  "Key": "Name",
-                  "Value": "SmokeTestApp/Vpc/Vpc"
-                }
-              ]
+    it('throws if there is not a subnet with a route pointing to a NAT gateway', async () => {
+      const mockVpc: ResourceDiffRecord = {
+        "stackName": "SmokeTestApp",
+        "format": IacFormat.awsCdk,
+        "changeType": ChangeType.CREATE,
+        "resourceType": "AWS::EC2::VPC",
+        "address": "Vpc/Vpc",
+        "logicalId": "VpcC3027511",
+        "properties": {
+          "cidrBlock": "10.40.0.0/16",
+          "instanceTenancy": "default",
+          "tagSet": [
+            {
+              "Key": "Name",
+              "Value": "SmokeTestApp/Vpc/Vpc"
             }
-          }
-        } as ResourceDiffRecord;
-        const allResources = require('../../../test-data/vpc-stack/CdkDiffWithNat.json');
-  
-        let thrownError;
-        try {
-          await vpcSmokeTest(mockVpc, allResources, { requirePrivateSubnet: true });
-        } catch (error) {
-          thrownError = error;
-        } finally {
-          expect(mockLoggerInfo).toBeCalled();
-          expect(mockLoggerInfo).toBeCalledWith('Verifying subnet configuration...');
-          expect(thrownError).toBeUndefined();
+          ]
         }
-      });
-      it('throws if there is not a subnet with a route pointing to a NAT gateway', async () => {
-        const mockVpc: ResourceDiffRecord = {
-          "stackName": "SmokeTestApp",
-          "format": IacFormat.awsCdk,
-          "changeType": ChangeType.CREATE,
-          "resourceType": "AWS::EC2::VPC",
-          "resourceRecord": {
-            "address": "Vpc/Vpc",
-            "type": "AWS::EC2::VPC",
-            "logicalId": "VpcC3027511",
-            "properties": {
-              "CidrBlock": "10.40.0.0/16",
-              "EnableDnsHostnames": true,
-              "EnableDnsSupport": true,
-              "InstanceTenancy": "default",
-              "Tags": [
-                {
-                  "Key": "Name",
-                  "Value": "SmokeTestApp/Vpc/Vpc"
-                }
-              ]
-            }
-          }
-        };
-        const allResources = require('../../../test-data/vpc-stack/CdkDiffNoNat.json');
-  
-        let thrownError;
-        try {
-          await vpcSmokeTest(mockVpc, allResources, { requirePrivateSubnet: true });
-        } catch (error) {
-          thrownError = error;
-        } finally {
-          expect(mockLoggerInfo).toBeCalled();
-          expect(mockLoggerInfo).toBeCalledWith('Verifying subnet configuration...');
-          
-          expect(thrownError).toBeDefined();
-          expect(thrownError).toHaveProperty('name', 'CustomError');
-          expect(thrownError).toHaveProperty('message', 'Missing private subnets!');
-          expect(thrownError).toHaveProperty('reason', 'Based on the configuration passed, private subnets with a NAT Gateway are required for all vpcs but none was found for "VpcC3027511".');
-        }
-      });
-    })
+      };
+      const allResources = require('../../../test-data/vpc-stack/cdk/no-nat/aws-cdk-diff.json');
+
+      let thrownError;
+      try {
+        await vpcSmokeTest(mockVpc, allResources, { requirePrivateSubnet: true });
+      } catch (error) {
+        thrownError = error;
+      } finally {
+        expect(mockLoggerInfo).toBeCalled();
+        expect(mockLoggerInfo).toBeCalledWith('Verifying subnet configuration...');
+        
+        expect(thrownError).toBeDefined();
+        expect(thrownError).toHaveProperty('name', 'CustomError');
+        expect(thrownError).toHaveProperty('message', 'Missing private subnets!');
+        expect(thrownError).toHaveProperty('reason', 'Based on the configuration passed, private subnets with a NAT Gateway are required for all vpcs but none was found for "VpcC3027511".');
+      }
+    });
   });
 });
