+++ test/cli/commands/smoke-test/test-data/vpc-stack/cdk/no-nat/diff.txt
@@ -1,0 +1,43 @@
+Stack SmokeTestApp
+
+Parameters
+
+[+] Parameter BootstrapVersion BootstrapVersion: {"Type":"AWS::SSM::Parameter::Value<String>","Default":"/cdk-bootstrap/hnb659fds/version","Description":"Version of the CDK Bootstrap resources in this environment, automatically retrieved from SSM Parameter Store. [cdk:skip]"}
+
+Conditions
+[+] Condition CDKMetadata/Condition CDKMetadataAvailable: {"Fn::Or":[{"Fn::Or":[{"Fn::Equals":[{"Ref":"AWS::Region"},"af-south-1"]},{"Fn::Equals":[{"Ref":"AWS::Region"},"ap-east-1"]},{"Fn::Equals":[{"Ref":"AWS::Region"},"ap-northeast-1"]},{"Fn::Equals":[{"Ref":"AWS::Region"},"ap-northeast-2"]},{"Fn::Equals":[{"Ref":"AWS::Region"},"ap-south-1"]},{"Fn::Equals":[{"Ref":"AWS::Region"},"ap-southeast-1"]},{"Fn::Equals":[{"Ref":"AWS::Region"},"ap-southeast-2"]},{"Fn::Equals":[{"Ref":"AWS::Region"},"ca-central-1"]},{"Fn::Equals":[{"Ref":"AWS::Region"},"cn-north-1"]},{"Fn::Equals":[{"Ref":"AWS::Region"},"cn-northwest-1"]}]},{"Fn::Or":[{"Fn::Equals":[{"Ref":"AWS::Region"},"eu-central-1"]},{"Fn::Equals":[{"Ref":"AWS::Region"},"eu-north-1"]},{"Fn::Equals":[{"Ref":"AWS::Region"},"eu-south-1"]},{"Fn::Equals":[{"Ref":"AWS::Region"},"eu-west-1"]},{"Fn::Equals":[{"Ref":"AWS::Region"},"eu-west-2"]},{"Fn::Equals":[{"Ref":"AWS::Region"},"eu-west-3"]},{"Fn::Equals":[{"Ref":"AWS::Region"},"me-south-1"]},{"Fn::Equals":[{"Ref":"AWS::Region"},"sa-east-1"]},{"Fn::Equals":[{"Ref":"AWS::Region"},"us-east-1"]},{"Fn::Equals":[{"Ref":"AWS::Region"},"us-east-2"]}]},{"Fn::Or":[{"Fn::Equals":[{"Ref":"AWS::Region"},"us-west-1"]},{"Fn::Equals":[{"Ref":"AWS::Region"},"us-west-2"]}]}]}
+
+Resources
+
+[+] AWS::EC2::VPC Vpc/Vpc VpcC3027511 
+
+[+] AWS::EC2::Subnet Vpc/Vpc/PublicSubnetSubnet1/Subnet VpcPublicSubnetSubnet1SubnetCE01ADA6 
+[+] AWS::EC2::RouteTable Vpc/Vpc/PublicSubnetSubnet1/RouteTable VpcPublicSubnetSubnet1RouteTableD16E6674 
+
+[+] AWS::EC2::SubnetRouteTableAssociation Vpc/Vpc/PublicSubnetSubnet1/RouteTableAssociation VpcPublicSubnetSubnet1RouteTableAssociationAF3B2E14 
+
+[+] AWS::EC2::Route Vpc/Vpc/PublicSubnetSubnet1/DefaultRoute VpcPublicSubnetSubnet1DefaultRoute0E1DFC20 
+[+] AWS::EC2::Subnet Vpc/Vpc/PublicSubnetSubnet2/Subnet VpcPublicSubnetSubnet2Subnet1C0D5211 
+[+] AWS::EC2::RouteTable Vpc/Vpc/PublicSubnetSubnet2/RouteTable VpcPublicSubnetSubnet2RouteTableD12B2355 
+[+] AWS::EC2::SubnetRouteTableAssociation Vpc/Vpc/PublicSubnetSubnet2/RouteTableAssociation VpcPublicSubnetSubnet2RouteTableAssociation31AD1654 
+
+[+] AWS::EC2::Route Vpc/Vpc/PublicSubnetSubnet2/DefaultRoute VpcPublicSubnetSubnet2DefaultRoute902DF4E4 
+[+] AWS::EC2::Subnet Vpc/Vpc/IsolatedSubnetSubnet1/Subnet VpcIsolatedSubnetSubnet1Subnet9413A915 
+[+] AWS::EC2::RouteTable Vpc/Vpc/IsolatedSubnetSubnet1/RouteTable VpcIsolatedSubnetSubnet1RouteTableE8228AAB 
+
+[+] AWS::EC2::SubnetRouteTableAssociation Vpc/Vpc/IsolatedSubnetSubnet1/RouteTableAssociation VpcIsolatedSubnetSubnet1RouteTableAssociationB0977E01 
+
+[+] AWS::EC2::Subnet Vpc/Vpc/IsolatedSubnetSubnet2/Subnet VpcIsolatedSubnetSubnet2Subnet68825D18 
+[+] AWS::EC2::RouteTable Vpc/Vpc/IsolatedSubnetSubnet2/RouteTable VpcIsolatedSubnetSubnet2RouteTable4C59DEE4 
+[+] AWS::EC2::SubnetRouteTableAssociation Vpc/Vpc/IsolatedSubnetSubnet2/RouteTableAssociation VpcIsolatedSubnetSubnet2RouteTableAssociationF1972EDB 
+[+] AWS::EC2::InternetGateway Vpc/Vpc/IGW VpcIGW488B0FEB 
+[+] AWS::EC2::VPCGatewayAttachment Vpc/Vpc/VPCGW VpcVPCGW42EC8516 
+
+Outputs
+[+] Output Vpc/VpcId VpcVpcId87CD4CDE: {"Description":"Vpc-vpc-id","Value":{"Ref":"VpcC3027511"}}
+
+Other Changes
+[+] Unknown Rules: {"CheckBootstrapVersion":{"Assertions":[{"Assert":{"Fn::Not":[{"Fn::Contains":[["1","2","3","4","5"],{"Ref":"BootstrapVersion"}]}]},"AssertDescription":"CDK bootstrap stack version 6 required. Please run 'cdk bootstrap' with a recent version of the CDK CLI."}]}}
+
+
+
