+++ test/cli/commands/smoke-test/test-data/vpc-stack/tf/no-nat/tf-diff.json
@@ -1,0 +1,131 @@
+[
+  {
+    "format": "tf",
+    "resourceType": "aws_internet_gateway",
+    "changeType": "CREATE",
+    "address": "aws_internet_gateway.ts_aws_internet_gateway",
+    "logicalId": "ts_aws_internet_gateway",
+    "providerName": "registry.terraform.io/hashicorp/aws"
+  },
+  {
+    "format": "tf",
+    "resourceType": "aws_route",
+    "changeType": "CREATE",
+    "address": "aws_route.ts_aws_route_public_igw",
+    "logicalId": "ts_aws_route_public_igw",
+    "providerName": "registry.terraform.io/hashicorp/aws",
+    "properties": {
+      "destinationCidrBlock": "0.0.0.0/0",
+      "gatewayId": "${aws_internet_gateway.ts_aws_internet_gateway.id}"
+    }
+  },
+  {
+    "format": "tf",
+    "resourceType": "aws_route_table",
+    "changeType": "CREATE",
+    "address": "aws_route_table.ts_aws_route_table_private_isolated",
+    "logicalId": "ts_aws_route_table_private_isolated",
+    "providerName": "registry.terraform.io/hashicorp/aws",
+    "properties": {
+      "associationSet": [
+        {
+          "routeTableId": "${aws_route_table.ts_aws_route_table_private_isolated.id}",
+          "subnetId": "${aws_subnet.ts_aws_subnet_private_isolated.id}"
+        }
+      ],
+      "routeSet": [],
+      "vpcId": "${aws_vpc.ts_aws_vpc.id}"
+    }
+  },
+  {
+    "format": "tf",
+    "resourceType": "aws_route_table",
+    "changeType": "CREATE",
+    "address": "aws_route_table.ts_aws_route_table_public_igw",
+    "logicalId": "ts_aws_route_table_public_igw",
+    "providerName": "registry.terraform.io/hashicorp/aws",
+    "properties": {
+      "associationSet": [
+        {
+          "routeTableId": "${aws_route_table.ts_aws_route_table_public_igw.id}",
+          "subnetId": "${aws_subnet.ts_aws_subnet_public_igw.id}"
+        }
+      ],
+      "routeSet": [
+        {
+          "destinationCidrBlock": "0.0.0.0/0",
+          "gatewayId": "${aws_internet_gateway.ts_aws_internet_gateway.id}"
+        }
+      ],
+      "vpcId": "${aws_vpc.ts_aws_vpc.id}"
+    }
+  },
+  {
+    "format": "tf",
+    "resourceType": "aws_route_table_association",
+    "changeType": "CREATE",
+    "address": "aws_route_table_association.ts_aws_route_table_association_private_isolated",
+    "logicalId": "ts_aws_route_table_association_private_isolated",
+    "providerName": "registry.terraform.io/hashicorp/aws",
+    "properties": {
+      "routeTableId": "${aws_route_table.ts_aws_route_table_private_isolated.id}",
+      "subnetId": "${aws_subnet.ts_aws_subnet_private_isolated.id}"
+    }
+  },
+  {
+    "format": "tf",
+    "resourceType": "aws_route_table_association",
+    "changeType": "CREATE",
+    "address": "aws_route_table_association.ts_aws_route_table_association_public_igw",
+    "logicalId": "ts_aws_route_table_association_public_igw",
+    "providerName": "registry.terraform.io/hashicorp/aws",
+    "properties": {
+      "routeTableId": "${aws_route_table.ts_aws_route_table_public_igw.id}",
+      "subnetId": "${aws_subnet.ts_aws_subnet_public_igw.id}"
+    }
+  },
+  {
+    "format": "tf",
+    "resourceType": "aws_subnet",
+    "changeType": "CREATE",
+    "address": "aws_subnet.ts_aws_subnet_private_isolated",
+    "logicalId": "ts_aws_subnet_private_isolated",
+    "providerName": "registry.terraform.io/hashicorp/aws",
+    "properties": {
+      "assignIpv6AddressOnCreation": false,
+      "cidrBlock": "10.0.48.0/20",
+      "enableDns64": false,
+      "ipv6Native": false,
+      "mapPublicIpOnLaunch": false,
+      "vpcId": "${aws_vpc.ts_aws_vpc.id}"
+    }
+  },
+  {
+    "format": "tf",
+    "resourceType": "aws_subnet",
+    "changeType": "CREATE",
+    "address": "aws_subnet.ts_aws_subnet_public_igw",
+    "logicalId": "ts_aws_subnet_public_igw",
+    "providerName": "registry.terraform.io/hashicorp/aws",
+    "properties": {
+      "assignIpv6AddressOnCreation": false,
+      "cidrBlock": "10.0.16.0/20",
+      "enableDns64": false,
+      "ipv6Native": false,
+      "mapPublicIpOnLaunch": true,
+      "vpcId": "${aws_vpc.ts_aws_vpc.id}"
+    }
+  },
+  {
+    "format": "tf",
+    "resourceType": "aws_vpc",
+    "changeType": "CREATE",
+    "address": "aws_vpc.ts_aws_vpc",
+    "logicalId": "ts_aws_vpc",
+    "providerName": "registry.terraform.io/hashicorp/aws",
+    "properties": {
+      "cidrBlock": "10.0.0.0/16",
+      "instanceTenancy": "default"
+    }
+  }
+]
