+++ test/cli/commands/smoke-test/test-data/vpc-stack/tf/with-nat/tf-json.json
@@ -1,0 +1,117 @@
+[
+  {
+    "name": "direct-main.tf",
+    "contents": {
+      "resource": {
+        "aws_eip": {
+          "ts_aws_eip_nat": [
+            {
+              "vpc": true
+            }
+          ]
+        },
+        "aws_internet_gateway": {
+          "ts_aws_internet_gateway": [
+            {
+              "vpc_id": "${aws_vpc.ts_aws_vpc.id}"
+            }
+          ]
+        },
+        "aws_nat_gateway": {
+          "ts_aws_nat_gateway": [
+            {
+              "allocation_id": "${aws_eip.ts_aws_eip_nat.id}",
+              "depends_on": [
+                "${aws_internet_gateway.ts_aws_internet_gateway}"
+              ],
+              "subnet_id": "${aws_subnet.ts_aws_subnet_public_igw.id}"
+            }
+          ]
+        },
+        "aws_route": {
+          "ts_aws_route_private_ngw": [
+            {
+              "destination_cidr_block": "0.0.0.0/0",
+              "nat_gateway_id": "${aws_nat_gateway.ts_aws_nat_gateway.id}",
+              "route_table_id": "${aws_route_table.ts_aws_route_table_private_ngw.id}"
+            }
+          ],
+          "ts_aws_route_public_igw": [
+            {
+              "destination_cidr_block": "0.0.0.0/0",
+              "gateway_id": "${aws_internet_gateway.ts_aws_internet_gateway.id}",
+              "route_table_id": "${aws_route_table.ts_aws_route_table_public_igw.id}"
+            }
+          ]
+        },
+        "aws_route_table": {
+          "ts_aws_route_table_private_isolated": [
+            {
+              "vpc_id": "${aws_vpc.ts_aws_vpc.id}"
+            }
+          ],
+          "ts_aws_route_table_private_ngw": [
+            {
+              "vpc_id": "${aws_vpc.ts_aws_vpc.id}"
+            }
+          ],
+          "ts_aws_route_table_public_igw": [
+            {
+              "vpc_id": "${aws_vpc.ts_aws_vpc.id}"
+            }
+          ]
+        },
+        "aws_route_table_association": {
+          "ts_aws_route_table_association_private_isolated": [
+            {
+              "route_table_id": "${aws_route_table.ts_aws_route_table_private_isolated.id}",
+              "subnet_id": "${aws_subnet.ts_aws_subnet_private_isolated.id}"
+            }
+          ],
+          "ts_aws_route_table_association_private_ngw": [
+            {
+              "route_table_id": "${aws_route_table.ts_aws_route_table_private_ngw.id}",
+              "subnet_id": "${aws_subnet.ts_aws_subnet_private_ngw.id}"
+            }
+          ],
+          "ts_aws_route_table_association_public_igw": [
+            {
+              "route_table_id": "${aws_route_table.ts_aws_route_table_public_igw.id}",
+              "subnet_id": "${aws_subnet.ts_aws_subnet_public_igw.id}"
+            }
+          ]
+        },
+        "aws_subnet": {
+          "ts_aws_subnet_private_isolated": [
+            {
+              "cidr_block": "${cidrsubnet(aws_vpc.ts_aws_vpc.cidr_block, var.ts_aws_vpc_cidr_newbits, 3)}",
+              "vpc_id": "${aws_vpc.ts_aws_vpc.id}"
+            }
+          ],
+          "ts_aws_subnet_private_ngw": [
+            {
+              "cidr_block": "${cidrsubnet(aws_vpc.ts_aws_vpc.cidr_block, var.ts_aws_vpc_cidr_newbits, 2)}",
+              "vpc_id": "${aws_vpc.ts_aws_vpc.id}"
+            }
+          ],
+          "ts_aws_subnet_public_igw": [
+            {
+              "cidr_block": "${cidrsubnet(aws_vpc.ts_aws_vpc.cidr_block, var.ts_aws_vpc_cidr_newbits, 1)}",
+              "map_public_ip_on_launch": true,
+              "vpc_id": "${aws_vpc.ts_aws_vpc.id}"
+            }
+          ]
+        },
+        "aws_vpc": {
+          "ts_aws_vpc": [
+            {
+              "cidr_block": "${var.ts_aws_vpc_cidr_block}",
+              "enable_dns_hostnames": true,
+              "instance_tenancy": "default"
+            }
+          ]
+        }
+      }
+    }
+  }
+]
